media: smfc: fix prevent issue
This patch fixed below prevent issue about other violation types.

CID 201997, 201944, 200890, 200853

(cherry picked from commit fed913ab79356dbd573144c8e63f868a3b4003fd)

+++ drivers/media/platform/exynos/smfc/smfc-regs.c
@@ -411,7 +411,7 @@ void smfc_hwconfigure_image(struct smfc_ctx *ctx,
 	stream_address = smfc_hwconfigure_jpeg_base(ctx, &vb2buf_jpg->vb2_buf,
 						    ctx->offset_of_sos, false);
 	if (!(ctx->flags & SMFC_CTX_COMPRESS)) {
-		u32 streamsize = vb2_plane_size(&vb2buf_jpg->vb2_buf, 0);
+		u32 streamsize = (u32)vb2_plane_size(&vb2buf_jpg->vb2_buf, 0);
 
 		streamsize -= ctx->offset_of_sos;
 		streamsize += stream_address & SMFC_ADDR_ALIGN_MASK(burstlen);
